SCRAPER/requirements.txt
appdirs==1.4.4
apturl==0.5.2
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configparser==5.0.2
crayons==0.4.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.12
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.12
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.2
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGithub==1.54.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.2.4
ssh-import-id==5.10
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.2
virtualenv-clone==0.3.0
wadllib==1.3.3
webdriver-manager==3.3.0
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
